+++ src/svelte.d.ts
@@ -3,3 +3,5 @@ declare module '*.svelte' {
   const component: Component
   export default component
 }
+
+declare module '*.css'